Basic support for DataContext={resource: ...}
This functionality is essential for server-side
only rendering, but this patch is only the groundwork -
we still need to add support to Repeater and possibly other controls.

The design is fairly simple - we allow resource binding
in data context and track that this data context does
not exist client-side (see the new ServerSideOnly property).
Using this server-only data context is not allowed in value
and staticCommand bindings. Note that using a value binding
inside the resource-bound control is allowed, it just cannot
reference `_this`. Since the context does not exist client-side,
we then translate _parent into $data instead of $parent.